fix: amend search result component to use suspense query
